+++ Source/Scene/SceneTransforms.js
@@ -82,7 +82,7 @@ define([
         var viewProjection = scene.getUniformState().getViewProjection();
         viewProjection.multiplyByPoint(actualPosition, positionCC);
 
-        return SceneTransforms.clipToWindowCoordinates(scene.getCanvas(), positionCC, result);
+        return SceneTransforms.clipToWindowCoordinates(scene.getContext(), positionCC, result);
     };
 
     var projectedPosition = new Cartesian3();
@@ -132,15 +132,15 @@ define([
     /**
      * @private
      */
-    SceneTransforms.clipToWindowCoordinates = function(canvas, position, result) {
+    SceneTransforms.clipToWindowCoordinates = function(context, position, result) {
         // Perspective divide to transform from clip coordinates to normalized device coordinates
         positionNDC.x = position.x / position.w;
         positionNDC.y = position.y / position.w;
         positionNDC.z = position.z / position.w;
 
         // Assuming viewport takes up the entire canvas...
-        viewport.width = canvas.clientWidth;
-        viewport.height = canvas.clientHeight;
+        viewport.width = context.getDrawingBufferWidth();
+        viewport.height = context.getDrawingBufferHeight();
         Matrix4.computeViewportTransformation(viewport, 0.0, 1.0, viewportTransform);
 
         // Viewport transform to transform from clip coordinates to window coordinates